explicit convert to uint32

+++ Source/Diagnostic/Time.c
@@ -80,8 +80,8 @@ void GetTimestamp(timestamp_t* storage, timestamp_format_t format)
 
     if (millisecond_count >= 1000)
     {
-        storage->seconds = millisecond_count / 1000;
-        storage->milliseconds = millisecond_count % 1000;
+        storage->seconds = (uint32_t)(millisecond_count / 1000);
+        storage->milliseconds = (uint32_t)(millisecond_count % 1000);
     }
     if (storage->seconds >= 60)
     {
